+++ src/integrationTest/java/uk/gov/hmcts/darts/transcriptions/controller/TranscriptionControllerUpdateTranscriptionWithTranscriberIntTest.java
@@ -59,7 +59,7 @@ class TranscriptionControllerUpdateTranscriptionWithTranscriberIntTest extends I
     private TranscriptionEntity transcriptionEntity;
     private UserAccountEntity testUser;
 
-    private Integer transcriptionId;
+    private Long transcriptionId;
     private Integer testUserId;
 
     @BeforeEach
