+++ src/main/java/uk/gov/hmcts/darts/audio/service/AudioPreviewService.java
@@ -4,5 +4,5 @@ import uk.gov.hmcts.darts.audio.model.AudioPreview;
 
 @FunctionalInterface
 public interface AudioPreviewService {
-    AudioPreview getOrCreateAudioPreview(Integer mediaId);
+    AudioPreview getOrCreateAudioPreview(Long mediaId);
 }
